# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "EZ Admin 后台管理系统 API，包含认证、用户管理、角色管理、系统配置等功能。",
        "title": "EZ Admin Gin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/system/notices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System / 公告管理"
                ],
                "summary": "查询公告列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ez-admin-gin_server_internal_modules_system_notice_api.ListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System / 公告管理"
                ],
                "summary": "创建公告",
                "parameters": [
                    {
                        "description": "公告参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ez-admin-gin_server_internal_modules_system_notice_api.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ez-admin-gin_server_internal_modules_system_notice_api.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    }
                }
            }
        },
        "/system/notices/{id}/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System / 公告管理"
                ],
                "summary": "删除公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    }
                }
            }
        },
        "/system/notices/{id}/status": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System / 公告管理"
                ],
                "summary": "更新公告状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ez-admin-gin_server_internal_modules_system_notice_api.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    }
                }
            }
        },
        "/system/notices/{id}/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System / 公告管理"
                ],
                "summary": "更新公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "公告参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ez-admin-gin_server_internal_modules_system_notice_api.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ez-admin-gin_server_internal_modules_system_notice_api.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errorsx.Code": {
            "type": "integer",
            "enum": [
                0,
                40000,
                40100,
                40300,
                40400,
                42900,
                50300,
                50000
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeTooManyRequests",
                "CodeServiceUnavailable",
                "CodeInternal"
            ]
        },
...
        "ez-admin-gin_server_internal_modules_system_notice_api.CreateRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.NoticeStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ez-admin-gin_server_internal_modules_system_notice_api.ListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ez-admin-gin_server_internal_modules_system_notice_domain.Response"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ez-admin-gin_server_internal_modules_system_notice_api.Response": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.NoticeStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "ez-admin-gin_server_internal_modules_system_notice_api.UpdateRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.NoticeStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ez-admin-gin_server_internal_modules_system_notice_api.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "$ref": "#/definitions/model.NoticeStatus"
                }
            }
        },
        "ez-admin-gin_server_internal_modules_system_notice_domain.Response": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.NoticeStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "httpx.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errorsx.Code"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.NoticeStatus": {
            "type": "integer",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "NoticeStatusEnabled",
                "NoticeStatusDisabled"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}